2025 Site v0.8
0.8 as in.. it's 80% of the way there!
There are still quite a few tweaks required - Both in logic and styling - but the overall base of the site is present.
Current projects are largely a port from the old site, and lack Repositories or articles (and translations!), but are suitable for getting the new site up and running.